Move KotlinDslVersion to org.gradle.util.internal.kotlin so it can be marked as NonNullApi
