+++ subprojects/launcher/src/main/java/org/gradle/launcher/cli/DefaultCommandLineActionFactory.java
@@ -48,7 +48,7 @@ import org.gradle.launcher.configuration.AllProperties;
 import org.gradle.launcher.configuration.BuildLayoutResult;
 import org.gradle.launcher.configuration.InitialProperties;
 import org.gradle.util.internal.DefaultGradleVersion;
-import org.gradle.util.internal.KotlinDslVersion;
+import org.gradle.util.internal.kotlin.KotlinDslVersion;
 
 import javax.annotation.Nullable;
 import java.io.PrintStream;
