+++ subprojects/logging/src/test/groovy/org/gradle/util/internal/kotlin/KotlinDslVersionTest.groovy
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.util.internal.kotlin
+
+
+import spock.lang.Specification
+
+class KotlinDslVersionTest extends Specification {
+
+    def "kotlin version should be present"() {
+        expect:
+        KotlinDslVersion.current().kotlinVersion != null
+    }
+}
